Add Docker and DC to Dockerfile.base
