+++ docker/Dockerfile.base
@@ -33,6 +33,12 @@ COPY env/common.sh env/
 COPY env/debian env/debian
 RUN sudo apt-get update && APTARGS=-y env/debian/deps && sudo apt-get clean
 
+# Install docker and docker-compose
+COPY tools/install_docker install_docker
+RUN sudo ./install_docker
+RUN rm install_docker
+RUN sudo usermod -aG docker scion
+
 # Pip packages
 COPY env/pip3 env/pip3
 COPY env/pip2 env/pip2
