+++ tools/dc
@@ -1,0 +1,26 @@
+#!/bin/bash
+
+export HOME_DIR=${HOME_DIR:-$PWD}
+
+cmd_help() {
+    echo
+	cat <<-_EOF
+	Usage:
+	    $PROGRAM scion [docker-compose command]
+	        Run docker-compose command for scion services.
+	_EOF
+}
+
+cmd_scion() {
+	COMPOSE_FILE="gen/scion-dc.yml" docker-compose --no-ansi "$@"
+}
+
+PROGRAM="${0##*/}"
+COMMAND="$1"
+shift
+
+case "$COMMAND" in
+	scion)
+	    "cmd_$COMMAND" "$@" ;;
+	*)  cmd_help; exit 1 ;;
+esac
