+++ tools/install_docker
@@ -1,0 +1,30 @@
+#!/bin/bash
+
+set -ex
+
+apt update
+apt install apt-transport-https ca-certificates curl software-properties-common iptables build-essential -y
+curl -fsSL https://download.docker.com/linux/ubuntu/gpg | apt-key add -
+apt-key fingerprint 0EBFCD88
+add-apt-repository \
+   "deb [arch=amd64] https://download.docker.com/linux/ubuntu \
+   $(lsb_release -cs) \
+   stable"
+
+# Update docker to 18.06
+apt update && apt install docker-ce=18.06.0~ce~3-0~ubuntu -y
+
+# Update docker-compose to 1.22.0
+case "$(uname -m)" in
+    x86_64)
+        src=https://github.com/docker/compose/releases/download/1.22.0/docker-compose-Linux-x86_64
+        sum=f679a24b93f291c3bffaff340467494f388c0c251649d640e661d509db9d57e9
+        file=/usr/local/bin/docker-compose
+        ;;
+    *)
+        echo "ERROR: unsupported architecture '$(uname -m)'"
+        exit 1
+esac
+curl -L "$src" -o "$file"
+echo "$sum $file" | sha256sum -c -
+chmod +x "$file"
